Move notiffcation functions to own file.
